search: just move populateSearchIndex into backend code
